Need to check keys when paths are provided also.
